Fixes issue 102 about missing using statements in interface layer code.
Added tests and fixed behavior for custom attribute data. We were just using the custom attribute's Type parameters FullName, and this caused problems since in the type picker, types always have Fullname + AssemblyName, and therefore we were not properly recognizing them as the same type. We now consistently use the fullname+assemblyname for these custom attributes, and this also meant that we didn't need to hardcode the two additional types to add, since they are always present from properties (i.e. Notes) from the runtime interface and/or info interfaces (for ProductCodeGeneration.cs)